feat(Header): refactor folders structure
