+++ tests/users/domain/models/UserMother.ts
@@ -1,4 +1,4 @@
-import { User } from '../../../src/domain/User'
+import { User } from '../../../../src/users/domain/models/User'
 import { faker } from '@faker-js/faker'
 
 export class UserMother {
